+++ internal/api/api.go
@@ -21,17 +21,95 @@ func NewApiService(cardanoGraphQLService cardano_graphql.Service) nvla.DefaultAp
 	}
 }
 
-// Availability information about Cardano APIs
-func (s *ApiService) GetCardanoStatus(ctx context.Context) (nvla.ImplResponse, error) {
+// TODO
+// GetOrders - Your GET endpoint
+func (s *ApiService) GetOrders(ctx context.Context, productId string, marketId string, organizationId string, count string) (nvla.ImplResponse, error) {
+	// TODO - update GetOrders with the required logic for this service method.
+	// Add api_default_service.go to the .openapi-generator-ignore to avoid overwriting this service implementation when updating open api generation.
+
+	//TODO: Uncomment the next line to return response Response(200, []Order{}) or use other options such as http.Ok ...
+	//return Response(200, []Order{}), nil
+
+	return nvla.Response(http.StatusNotImplemented, nil), errors.New("GetOrders method not implemented")
+}
+
+// Gets listed products
+func (s *ApiService) GetProducts(ctx context.Context, marketId string, organizationId string, productId string) (nvla.ImplResponse, error) {
+	// TODO: unmock, use query params
+
+	products := []nvla.Product{
+		{
+			Asset: nvla.ProductAsset{
+				Urls: []string{
+					"https://siasky.net/AACQHdh6YsQfFshLfHjhjsQNCwKHbnaJ2CryhZRAs4HqNQ",
+					"https://api.rektangularstudios.com/ipfs/QmXR2TGqndCHmu4utjpzrSwaGtUVMvrjBQ8eKwpRAnPnTh",
+					"https://api.rektangularstudios.com/static/cards/iscara_the_ten_thousand_guns/iscara_the_ten_thousand_guns.zip",
+				},
+				Loader: "occulta_novellia_character",
+			},
+			Pricing: nvla.ProductPricing{
+				CurrencyPolicyId: "ada",
+				UnitPrice: 20,
+				MaxOrderSize: 5,
+				PurchaseAddress: "0xDraculiDepositAddress",
+				DateAvailable: "2021-05-15",
+			},
+			Organization: nvla.ProductOrganization{
+				Name: "Rektangular Studios",
+				OrganizationId: 1,
+			},
+			Market: nvla.ProductMarket{
+				Name: "Occulta Novellia",
+				MarketId: 1,
+			},
+			Stock: nvla.ProductStock{
+				Available: 2400,
+				TotalSupply: 2500,
+			},
+			Metadata: nvla.ProductMetadata{
+				Tags: []string{"Game Character"},
+				DateListed: "2021-05-01",
+			},
+			Artist: nvla.ProductArtist{
+				Name: "ArtistName",
+				Urls: []string{
+					"https://www.artstation.com/",
+					"https://www.deviantart.com/",
+				},
+			},
+			Product: nvla.ProductProduct{
+				OverviewImageUrls: []string{
+					"https://siasky.net/_Aqfe5wxnzn54sPn2cBPNnwmjQ2te4rXZfVNvUb79QhBWw",
+					"https://api.rektangularstudios.com/ipfs/QmQbju6V8vpjKS1AG9vw4mQucYoeiCY2dNSnrsWzfJAZue",
+					"https://api.rektangularstudios.com/static/cards/iscara_the_ten_thousand_guns/iscara_the_ten_thousand_guns_card.jpg",
+				},
+				TokenPolicyId: "0xrandomNumbers.IscaraTheTenThousandGuns",
+				DescriptionShort: "Occulta Novellia Character",
+				DescriptionLong: "A character token for the upcoming surreal horror game Occulta Novellia. This token will grant access to a playable character.",
+				Name: "Iscara the Ten Thousand Guns",
+				ProductId: 2,
+			},
+		},
+	}
+
+	return nvla.Response(200, products), nil
+}
+
+// Availability information about service availability
+func (s *ApiService) GetStatus(ctx context.Context) (nvla.ImplResponse, error) {
 	initialized, syncPercentage, err := s.cardanoGraphQLService.GetStatus(ctx)
 	
 	if err != nil {
 		return nvla.Response(500, fmt.Sprintf("error: %v", err)), nil
 	}
 
-	resp := nvla.CardanoStatus{
-		Initialized: initialized,
-		SyncPercentage: syncPercentage,
+	resp := nvla.Status{
+		Cardano: nvla.StatusCardano{
+			Initialized: initialized,
+			SyncPercentage: syncPercentage,
+		},
+		// TODO: read this value from somewhere
+		Maintenance: false,
 	}
 	return nvla.Response(200, resp), nil
 }